stt_service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
openai-whisper==20231117
torch==2.1.2
numpy==1.22.0
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
python-multipart==0.0.6
soundfile==0.12.1
redis==5.0.1
prometheus-client==0.19.0 